Improve TS implementations of BlazorUI #7374
