+++ src/BlazorUI/Bit.BlazorUI.Extras/Components/Chart/Interop/BitChartJsInterop.cs
@@ -22,12 +22,12 @@ internal static class BitChartJsInterop
 
     public static async ValueTask InitChartJs(this IJSRuntime jsRuntime, IEnumerable<string> scripts)
     {
-        await jsRuntime.InvokeVoidAsync($"BitChart.initChartJs", scripts);
+        await jsRuntime.InvokeVoidAsync($"BitBlazorUI.BitChart.initChartJs", scripts);
     }
 
     public static async ValueTask RemoveChart(this IJSRuntime jsRuntime, string canvasId)
     {
-        await jsRuntime.InvokeVoidAsync($"BitChart.removeChart", canvasId);
+        await jsRuntime.InvokeVoidAsync($"BitBlazorUI.BitChart.removeChart", canvasId);
     }
 
     /// <summary>
@@ -40,7 +40,7 @@ internal static class BitChartJsInterop
     {
         dynamic dynParam = StripNulls(chartConfig);
         Dictionary<string, object> param = ConvertExpandoObjectToDictionary(dynParam);
-        return jsRuntime.InvokeAsync<bool>($"BitChart.setupChart", param);
+        return jsRuntime.InvokeAsync<bool>($"BitBlazorUI.BitChart.setupChart", param);
     }
 
     /// <summary>
@@ -95,7 +95,7 @@ internal static class BitChartJsInterop
     {
         dynamic dynParam = StripNulls(chartConfig);
         Dictionary<string, object> param = ConvertExpandoObjectToDictionary(dynParam);
-        return jsRuntime.InvokeAsync<bool>($"BitChart.updateChart", param);
+        return jsRuntime.InvokeAsync<bool>($"BitBlazorUI.BitChart.updateChart", param);
     }
 
     /// <summary>
