+++ src/BlazorUI/Bit.BlazorUI.Extras/Components/DataGrid/BitDataGridJsInterop.cs
@@ -1,0 +1,14 @@
+﻿namespace Bit.BlazorUI;
+
+internal static class BitDataGridJsInterop
+{
+    public static async ValueTask<IJSObjectReference> BitDataGridInit(this IJSRuntime jsRuntime, ElementReference tableElement)
+    {
+        return await jsRuntime.InvokeAsync<IJSObjectReference>("BitBlazorUI.BitDataGrid.init", tableElement);
+    }
+
+    public static async ValueTask BitDataGridCheckColumnOptionsPosition(this IJSRuntime jsRuntime, ElementReference tableElement)
+    {
+        await jsRuntime.InvokeVoidAsync("BitBlazorUI.BitDataGrid.checkColumnOptionsPosition", tableElement);
+    }
+}
