+++ src/BlazorUI/Bit.BlazorUI/Components/Breadcrumb/BitBreadcrumbJsRuntimeExtension.cs
@@ -1,17 +1,21 @@
-﻿using Bit.BlazorUI;
-
-namespace Microsoft.JSInterop;
+﻿namespace Bit.BlazorUI;
 
 internal static class BitBreadcrumbJsRuntimeExtension
 {
-    internal static async Task ToggleOverflowCallout<T>(this IJSRuntime jsRuntime,
-                                                        DotNetObjectReference<BitBreadcrumb<T>> component,
-                                                        string wrapperId,
-                                                        string dropdownId,
-                                                        string calloutId,
-                                                        string overlayId,
-                                                        bool isCalloutOpen) where T : class
+    internal static async Task BitBreadcrumbToggleOverflowCallout<T>(this IJSRuntime jsRuntime,
+                                                                     DotNetObjectReference<BitBreadcrumb<T>> component,
+                                                                     string wrapperId,
+                                                                     string dropdownId,
+                                                                     string calloutId,
+                                                                     string overlayId,
+                                                                     bool isCalloutOpen) where T : class
     {
-        await jsRuntime.InvokeAsync<string>("BitBreadcrumb.toggleOverflowCallout", component, wrapperId, dropdownId, calloutId, overlayId, isCalloutOpen);
+        await jsRuntime.InvokeAsync<string>("BitBlazorUI.Breadcrumb.toggleOverflowCallout",
+                                            component,
+                                            wrapperId,
+                                            dropdownId,
+                                            calloutId,
+                                            overlayId,
+                                            isCalloutOpen);
     }
 }
