+++ src/BlazorUI/Bit.BlazorUI/Components/FileUpload/BitFileUploadJsExtension.cs
@@ -1,44 +1,43 @@
 ﻿using System.Diagnostics.CodeAnalysis;
-using Bit.BlazorUI;
 
-namespace Microsoft.JSInterop;
+namespace Bit.BlazorUI;
 
 internal static class BitFileUploadJsExtension
 {
-    internal static async Task<IJSObjectReference> SetupFileUploadDropzone(this IJSRuntime jsRuntime, ElementReference dragDropZoneElement, ElementReference inputFileElement)
-    {
-        return await jsRuntime.InvokeAsync<IJSObjectReference>("BitFileUpload.setupDropzone", dragDropZoneElement, inputFileElement);
-    }
-
     [DynamicDependency(DynamicallyAccessedMemberTypes.All, typeof(BitFileInfo))]
-    internal static async Task<BitFileInfo[]?> ResetFileUpload(this IJSRuntime jsRuntime,
+    internal static async Task<BitFileInfo[]?> BitFileUploadReset(this IJSRuntime jsRuntime,
                                                                    Guid id,
                                                                    DotNetObjectReference<BitFileUpload>? dotnetObjectReference,
                                                                    ElementReference element,
                                                                    string uploadAddress,
                                                                    IReadOnlyDictionary<string, string> uploadRequestHttpHeaders)
     {
-        return await jsRuntime.InvokeAsync<BitFileInfo[]>("BitFileUpload.reset", id.ToString(), dotnetObjectReference, element, uploadAddress, uploadRequestHttpHeaders);
+        return await jsRuntime.InvokeAsync<BitFileInfo[]>("BitBlazorUI.FileUpload.reset", id.ToString(), dotnetObjectReference, element, uploadAddress, uploadRequestHttpHeaders);
     }
 
-    internal static async Task UploadFile(this IJSRuntime jsRuntime, Guid id, long from, long to, int index, string? uploadUrl, IReadOnlyDictionary<string, string>? httpHeaders)
+    internal static async Task BitFileUploadUpload(this IJSRuntime jsRuntime, Guid id, long from, long to, int index, string? uploadUrl, IReadOnlyDictionary<string, string>? httpHeaders)
     {
-        await (httpHeaders is null ? jsRuntime.InvokeVoidAsync("BitFileUpload.upload", id.ToString(), from, to, index, uploadUrl)
-                                   : jsRuntime.InvokeVoidAsync("BitFileUpload.upload", id.ToString(), from, to, index, uploadUrl, httpHeaders));
+        await (httpHeaders is null ? jsRuntime.InvokeVoidAsync("BitBlazorUI.FileUpload.upload", id.ToString(), from, to, index, uploadUrl)
+                                   : jsRuntime.InvokeVoidAsync("BitBlazorUI.FileUpload.upload", id.ToString(), from, to, index, uploadUrl, httpHeaders));
     }
 
-    internal static async Task PauseFile(this IJSRuntime jsRuntime, Guid id, int index = -1)
+    internal static async Task BitFileUploadPause(this IJSRuntime jsRuntime, Guid id, int index = -1)
     {
-        await jsRuntime.InvokeVoidAsync("BitFileUpload.pause", id.ToString(), index);
+        await jsRuntime.InvokeVoidAsync("BitBlazorUI.FileUpload.pause", id.ToString(), index);
     }
 
-    internal static async Task BrowseFile(this IJSRuntime jsRuntime, ElementReference inputFileElement)
+    internal static async Task<IJSObjectReference> BitFileUploadSetupDragDrop(this IJSRuntime jsRuntime, ElementReference dragDropZoneElement, ElementReference inputFileElement)
     {
-        await jsRuntime.InvokeVoidAsync("BitFileUpload.browse", inputFileElement);
+        return await jsRuntime.InvokeAsync<IJSObjectReference>("BitBlazorUI.FileUpload.setupDragDrop", dragDropZoneElement, inputFileElement);
     }
 
-    internal static async Task DisposeFileUpload(this IJSRuntime jsRuntime, Guid id)
+    internal static async Task BitFileUploadBrowse(this IJSRuntime jsRuntime, ElementReference inputFileElement)
     {
-        await jsRuntime.InvokeVoidAsync("BitFileUpload.dispose", id.ToString());
+        await jsRuntime.InvokeVoidAsync("BitBlazorUI.FileUpload.browse", inputFileElement);
+    }
+
+    internal static async Task BitFileUploadDispose(this IJSRuntime jsRuntime, Guid id)
+    {
+        await jsRuntime.InvokeVoidAsync("BitBlazorUI.FileUpload.dispose", id.ToString());
     }
 }
