+++ src/BlazorUI/Bit.BlazorUI/Components/Inputs/OtpInput/BitOtpInput.ts
@@ -1,14 +1,15 @@
-﻿class BitOtpInput {
-    static setupOtpInput(dotnetReference: DotNetObject, otpInput: HTMLInputElement) {
+﻿namespace BitBlazorUI {
+    export class OtpInput {
+        public static setup(dotnetReference: DotNetObject, input: HTMLInputElement) {
+            input.addEventListener('focus', (e: any) => {
+                e.target?.select();
+            });
 
-        otpInput.addEventListener('focus', (e: any) => {
-            e.target?.select();
-        });
-
-        otpInput.addEventListener('paste', async e => {
-            e.preventDefault();
-            let pastedValue = e.clipboardData?.getData('Text');
-            await dotnetReference.invokeMethodAsync("SetPastedData", pastedValue);
-        });
+            input.addEventListener('paste', async e => {
+                e.preventDefault();
+                let pastedValue = e.clipboardData?.getData('Text');
+                await dotnetReference.invokeMethodAsync("SetPastedData", pastedValue);
+            });
+        }
     }
 }
