+++ src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/BitSearchBox.razor.cs
@@ -452,7 +452,7 @@ public partial class BitSearchBox
 
         CurrentValue = _searchItems[_selectedIndex];
         await OnChange.InvokeAsync(CurrentValue);
-        await _js.InvokeVoidAsync("BitSearchBox.moveCursorToEnd", _inputRef);
+        await _js.BitSearchBoxMoveCursorToEnd(_inputRef);
     }
 
     private async Task HandleOnItemClick(string item)
