+++ src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/BitSearchBox.ts
@@ -1,5 +1,7 @@
-﻿class BitSearchBox {
-    static moveCursorToEnd(inputElement: HTMLInputElement) {
-        inputElement.selectionStart = inputElement.selectionEnd = inputElement.value.length;
+﻿namespace BitBlazorUI {
+    export class SearchBox {
+        public static moveCursorToEnd(inputElement: HTMLInputElement) {
+            inputElement.selectionStart = inputElement.selectionEnd = inputElement.value.length;
+        }
     }
 }
