+++ src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/BitSearchBoxJsRuntimeExtension.cs
@@ -1,0 +1,9 @@
+﻿namespace Bit.BlazorUI;
+
+internal static class BitSearchBoxJsRuntimeExtension
+{
+    internal static async Task BitSearchBoxMoveCursorToEnd(this IJSRuntime jsRuntime, ElementReference input)
+    {
+        await jsRuntime.InvokeVoidAsync("BitBlazorUI.SearchBox.moveCursorToEnd", input);
+    }
+}
