+++ src/BlazorUI/Bit.BlazorUI/Components/Modal/BitModal.razor.cs
@@ -152,16 +152,16 @@ public partial class BitModal : IDisposable
         {
             if (IsDraggable)
             {
-                _ = _js.SetupDragDrop(_containerId, GetDragElementSelector());
+                _ = _js.BitModalSetupDragDrop(_containerId, GetDragElementSelector());
             }
             else
             {
-                _ = _js.RemoveDragDrop(_containerId, GetDragElementSelector());
+                _ = _js.BitModalRemoveDragDrop(_containerId, GetDragElementSelector());
             }
         }
         else
         {
-            _ = _js.RemoveDragDrop(_containerId, GetDragElementSelector());
+            _ = _js.BitModalRemoveDragDrop(_containerId, GetDragElementSelector());
         }
 
         _offsetTop = 0;
@@ -218,7 +218,7 @@ public partial class BitModal : IDisposable
     {
         if (_disposed || disposing is false) return;
 
-        _ = _js.RemoveDragDrop(_containerId, GetDragElementSelector());
+        _ = _js.BitModalRemoveDragDrop(_containerId, GetDragElementSelector());
 
         _disposed = true;
     }
