+++ src/BlazorUI/Bit.BlazorUI/Components/Modal/BitModal.ts
@@ -1,75 +1,77 @@
-﻿class BitModal {
-    private static DragDropListeners: any = {};
+﻿namespace BitBlazorUI {
+    export class Modal {
+        private static _dragDropListeners: any = {};
 
-    public static setupDragDrop(containerId: string, dragElementSelector: string) {
-        BitModal.removeDragDrop(containerId, dragElementSelector);
-        const listeners: any = {};
-        BitModal.DragDropListeners[containerId] = listeners;
+        public static setupDragDrop(containerId: string, dragElementSelector: string) {
+            Modal.removeDragDrop(containerId, dragElementSelector);
+            const listeners: any = {};
+            Modal._dragDropListeners[containerId] = listeners;
 
-        const element = document.getElementById(containerId)! as HTMLElement;
-        const dragElement = document.querySelector(dragElementSelector)! as HTMLElement;
+            const element = document.getElementById(containerId)! as HTMLElement;
+            const dragElement = document.querySelector(dragElementSelector)! as HTMLElement;
 
-        let x = 0;
-        let y = 0;
+            let x = 0;
+            let y = 0;
 
-        listeners['pointerdown'] = handlePointerDown;
-        dragElement.addEventListener('pointerdown', handlePointerDown);
-        dragElement.style.cursor = 'move';
-        dragElement.classList.add('bit-mdl-nta');
+            listeners['pointerdown'] = handlePointerDown;
+            dragElement.addEventListener('pointerdown', handlePointerDown);
+            dragElement.style.cursor = 'move';
+            dragElement.classList.add('bit-mdl-nta');
 
-        function handlePointerDown(e: PointerEvent) {
-            //e.preventDefault();
+            function handlePointerDown(e: PointerEvent) {
+                //e.preventDefault();
 
-            x = e.clientX;
-            y = e.clientY;
+                x = e.clientX;
+                y = e.clientY;
 
-            document.addEventListener('pointermove', handlePointerMove);
-            listeners['pointermove'] = handlePointerMove;
+                document.addEventListener('pointermove', handlePointerMove);
+                listeners['pointermove'] = handlePointerMove;
 
-            document.addEventListener('pointerup', handlePointerUp);
-            //document.addEventListener('pointerout', handlePointerUp);
-            //document.addEventListener('pointerleave', handlePointerUp);
-            listeners['pointerup'] = handlePointerUp;
+                document.addEventListener('pointerup', handlePointerUp);
+                //document.addEventListener('pointerout', handlePointerUp);
+                //document.addEventListener('pointerleave', handlePointerUp);
+                listeners['pointerup'] = handlePointerUp;
+            }
+
+            function handlePointerMove(e: PointerEvent) {
+                e.preventDefault();
+
+                element.style.left = `${element.offsetLeft - (x - e.clientX)}px`;
+                element.style.top = `${element.offsetTop - (y - e.clientY)}px`;
+
+                x = e.clientX;
+                y = e.clientY;
+            }
+
+            function handlePointerUp() {
+                document.removeEventListener('pointermove', handlePointerMove);
+
+                document.removeEventListener('pointerup', handlePointerUp);
+                //document.removeEventListener('pointerout', handlePointerUp);
+                //document.removeEventListener('pointerleave', handlePointerUp);
+            }
         }
 
-        function handlePointerMove(e: PointerEvent) {
-            e.preventDefault();
+        public static removeDragDrop(id: string, dragElementSelector: string) {
+            const listeners = Modal._dragDropListeners[id];
+            if (!listeners) return;
 
-            element.style.left = `${element.offsetLeft - (x - e.clientX)}px`;
-            element.style.top = `${element.offsetTop - (y - e.clientY)}px`;
+            const dragElement = document.querySelector(dragElementSelector)! as HTMLElement;
 
-            x = e.clientX;
-            y = e.clientY;
+            dragElement.removeEventListener('pointerdown', listeners['pointerdown']);
+            dragElement.style.cursor = '';
+            dragElement.classList.remove('bit-mdl-nta');
+
+            document.removeEventListener('pointermove', listeners['pointermove']);
+
+            document.removeEventListener('pointerup', listeners['pointerup']);
+            //document.removeEventListener('pointerout', listeners['pointerup']);
+            //document.removeEventListener('pointerleave', listeners['pointerup']);
+
+            delete listeners['pointerdown'];
+            delete listeners['pointermove'];
+            delete listeners['pointerup'];
+            delete Modal._dragDropListeners[id];
         }
-
-        function handlePointerUp() {
-            document.removeEventListener('pointermove', handlePointerMove);
-
-            document.removeEventListener('pointerup', handlePointerUp);
-            //document.removeEventListener('pointerout', handlePointerUp);
-            //document.removeEventListener('pointerleave', handlePointerUp);
-        }
-    }
-
-    public static removeDragDrop(id: string, dragElementSelector: string) {
-        const listeners = BitModal.DragDropListeners[id];
-        if (!listeners) return;
-
-        const dragElement = document.querySelector(dragElementSelector)! as HTMLElement;
-
-        dragElement.removeEventListener('pointerdown', listeners['pointerdown']);
-        dragElement.style.cursor = '';
-        dragElement.classList.remove('bit-mdl-nta');
-
-        document.removeEventListener('pointermove', listeners['pointermove']);
-
-        document.removeEventListener('pointerup', listeners['pointerup']);
-        //document.removeEventListener('pointerout', listeners['pointerup']);
-        //document.removeEventListener('pointerleave', listeners['pointerup']);
-
-        delete listeners['pointerdown'];
-        delete listeners['pointermove'];
-        delete listeners['pointerup'];
-        delete BitModal.DragDropListeners[id];
     }
 }
