+++ src/BlazorUI/Bit.BlazorUI/Components/Overlay/BitOverlay.ts
@@ -1,11 +1,13 @@
-﻿class BitOverlay {
-    public static toggleScroll(selector: string, isVisible: boolean) {
-        const element = document.querySelector(selector) as HTMLElement;
+﻿namespace BitBlazorUI {
+    export class Overlay {
+        public static toggleScroll(selector: string, isVisible: boolean) {
+            const element = document.querySelector(selector) as HTMLElement;
 
-        if (!element) return 0;
+            if (!element) return 0;
 
-        element.style.overflow = isVisible ? "hidden" : "auto";
+            element.style.overflow = isVisible ? "hidden" : "auto";
 
-        return element.scrollTop;
+            return element.scrollTop;
+        }
     }
 }
