+++ src/BlazorUI/Bit.BlazorUI/Components/Overlay/BitOverlayJsExtensions.cs
@@ -1,10 +1,9 @@
-﻿
-namespace Bit.BlazorUI;
+﻿namespace Bit.BlazorUI;
 
 internal static class BitOverlayJsExtensions
 {
-    internal static async Task<int> ToggleOverlayScroll(this IJSRuntime jsRuntime, string scrollerSelector, bool isVisible)
+    internal static async Task<int> BitOverlayToggleScroll(this IJSRuntime jsRuntime, string scrollerSelector, bool isVisible)
     {
-        return await jsRuntime.InvokeAsync<int>("BitOverlay.toggleScroll", scrollerSelector, isVisible);
+        return await jsRuntime.InvokeAsync<int>("BitBlazorUI.Overlay.toggleScroll", scrollerSelector, isVisible);
     }
 }
