+++ src/BlazorUI/Bit.BlazorUI/Components/Swiper/BitSwiper.ts
@@ -1,24 +1,26 @@
-﻿class BitSwiper {
-    static getDimensions(root: HTMLDivElement, swiper: HTMLDivElement) {
-        const swiperWidth = [].slice.call(swiper.children).reduce((pre, cur: HTMLDivElement) => pre + cur.offsetWidth, 0);
-        const effectiveSwiperWidth = swiperWidth - (swiper.parentElement?.offsetWidth ?? 0);
+﻿namespace BitBlazorUI {
+    export class Swiper {
+        public static getDimensions(root: HTMLDivElement, swiper: HTMLDivElement) {
+            const swiperWidth = [].slice.call(swiper.children).reduce((pre, cur: HTMLDivElement) => pre + cur.offsetWidth, 0);
+            const effectiveSwiperWidth = swiperWidth - (swiper.parentElement?.offsetWidth ?? 0);
 
-        const computedStyle = window.getComputedStyle(swiper);
-        const matrix = computedStyle.getPropertyValue('transform');
-        const matched = matrix.match(/matrix\((.+)\)/);
+            const computedStyle = window.getComputedStyle(swiper);
+            const matrix = computedStyle.getPropertyValue('transform');
+            const matched = matrix.match(/matrix\((.+)\)/);
 
-        let swiperTranslateX = 0;
-        if (matched && matched.length > 1) {
-            const splitted = matched[1].split(',');
-            swiperTranslateX = +splitted[4];
+            let swiperTranslateX = 0;
+            if (matched && matched.length > 1) {
+                const splitted = matched[1].split(',');
+                swiperTranslateX = +splitted[4];
+            }
+
+            return { rootWidth: root.offsetWidth, swiperWidth, effectiveSwiperWidth, swiperTranslateX };
         }
 
-        return { rootWidth: root.offsetWidth, swiperWidth, effectiveSwiperWidth, swiperTranslateX };
-    }
-
-    static registerPointerLeave(root: HTMLDivElement, dotnetObj: DotNetObject) {
-        root.addEventListener('pointerleave', async e => {
-            await dotnetObj.invokeMethodAsync("HandlePointerLeave", e.clientX);
-        });
+        public static registerPointerLeave(root: HTMLDivElement, dotnetObj: DotNetObject) {
+            root.addEventListener('pointerleave', async e => {
+                await dotnetObj.invokeMethodAsync("HandlePointerLeave", e.clientX);
+            });
+        }
     }
 }
