+++ src/BlazorUI/Bit.BlazorUI/Scripts/general.ts
@@ -4,23 +4,19 @@
     dispose(): void;
 }
 
-//window.addEventListener('load', e => {
-//    Bit.init();
-//})
-
 window.addEventListener('scroll', (e: Event) => {
-    const currentCallout = BitCallouts.current;
-    if (window.innerWidth < Bit.MAX_MOBILE_WIDTH && currentCallout.responsiveMode) return;
+    const currentCallout = BitBlazorUI.Callouts.current;
+    if (window.innerWidth < BitBlazorUI.Utils.MAX_MOBILE_WIDTH && currentCallout.responsiveMode) return;
 
     const target = e.target as HTMLElement;
     if (target?.id && target.id == currentCallout.scrollContainerId) return;
 
-    BitCallouts.replaceCurrent();
+    BitBlazorUI.Callouts.replaceCurrent();
 }, true);
 
 window.addEventListener('resize', (e: any) => {
     const resizeTriggeredByOpenningKeyboard = document?.activeElement?.getAttribute('type') === 'text';
-    if (window.innerWidth < Bit.MAX_MOBILE_WIDTH && resizeTriggeredByOpenningKeyboard) return;
+    if (window.innerWidth < BitBlazorUI.Utils.MAX_MOBILE_WIDTH && resizeTriggeredByOpenningKeyboard) return;
 
-    BitCallouts.replaceCurrent();
+    BitBlazorUI.Callouts.replaceCurrent();
 }, true);
