+++ src/BlazorUI/Demo/Client/Bit.BlazorUI.Demo.Client.Core/Pages/Components/Modal/BitModalDemo.razor
@@ -235,7 +235,12 @@
                         </p>
                     </div>
                 </BitModal>
+                <br />
+                <br />
+                <br />
                 <hr style="border-color:white" />
+                <br />
+                <br />
                 <div class="example-desc">Using custom drag element.</div>
                 <BitButton OnClick="() => IsOpen6 = true">Open Modal</BitButton>
                 <BitModal @bind-IsOpen="IsOpen6" IsDraggable="true" DragElementSelector=".modal-header-drag">
